requirements.txt
Django==5.0.7
numpy==2.1.0
Pillow==10.1.0
Pillow==10.4.0
tensorflow==2.17.0
tensorflow_intel==2.17.0
